Add transaction support for cache operations
- Speeds up course configuring as the invalidation is done in a single
access to the cache.
- Prevents error situations where cache was generated between the
invalidation being called and the transaction being committed. In such
cases, the generated item would overwrite the invalidation but still use
old data as the transaction hasn't been committed to the database.
- Prevents unnecessary cache invalidation in the case that the transaction
is rolled back.
- Prevents non-committed transaction statements from affecting other
requests. Currently, this was only an issue in tests because that is the
only place that generates cache items inside a transaction.