Use cached content in invalidation
- the content cache is cached in-memory inside a request
global for subsequent invalidations
- makes multiple invalidations much faster

The caching of the content cache can technically result in
invalidation not working correctly. However, this requires a very
specific order of events: (1) the content cache is cached in-memory,
(2) (either by the current or another request) the parent/module of
an exercise is modified in the database and the content cache is
regenerated, and (3) current request (the one from (1)) invalidates
the content cache for the modified exercise.